Update ConfigFileService and add more sections
This commit incorporates IAcsiConfigFileSection and IScsiConfigFileSection to ConfigFileService for more versatility. It also restructures ConfigFileSection and various related interfaces and classes into separate files for better organization and readability. It introduces amendments in the UI which involve label position modifications and incorporation of more user options in features like byte swapping.