Add initial selection logic in TimerTick method
The 'TimerTick' method has been updated to include logic for initial selection of a configuration in the 'Systems' list. This extra logic helps to avoid null selection scenarios on windows. This step was mainly necessary due to a current issue with Maui affecting the initial selection process on Windows, but the changes might be reverted once the issue is resolved by Maui.