Add feature to open current folder in explorer
The commit adds a new command in the FujiFilePickerPopupViewModel that allows opening the current folder in the file explorer. It also modifies FilePickerBreadCrumb.xaml to add a button that triggers this new command. This change provides a quicker way for users to open the current folder directly from the application.